Migrate slider from /docs/components/slider to /components/slider
Move all examples (Basic, Form, Custom Range) from the old documentation
page to the new reference page with Props Playground. Remove the old page,
update routes, sidebar navigation, and E2E tests.
